refactor(http): retire residual composition forwarding
