+++ pkg/services/factory_sessions/transports/http/handler.go
@@ -112,6 +112,48 @@ func NewHandler(deps Dependencies, logger *zap.Logger) *Adapter {
 	}
 }
 
+// FactoryStatusSessionReader is the narrow session-scoped observation role
+// used by the Factory Sessions status routes. Wire supplies the owner root
+// only after confirming it exposes this capability.
+type FactoryStatusSessionReader interface {
+	ObserveForSession(
+		context.Context,
+		string,
+		factoryruntime.ObserveRequest,
+	) (factoryruntime.ObserveResult, error)
+}
+
+type factoryStatusAPI struct {
+	sessions  FactoryStatusSessionReader
+	projector factoryruntime.FactoryStatusProjector
+}
+
+// NewFactoryStatusAPI binds status projection to the Factory Sessions session
+// router. Factory Sessions owns session identity; the selected session gateway
+// owns the live observation.
+func NewFactoryStatusAPI(
+	sessions FactoryStatusSessionReader,
+	projector factoryruntime.FactoryStatusProjector,
+) apisurface.FactoryStatusAPI {
+	return &factoryStatusAPI{sessions: sessions, projector: projector}
+}
+
+func (api *factoryStatusAPI) ProjectFactoryStatus(ctx context.Context, sessionID string) (factoryruntime.FactoryStatus, error) {
+	if api == nil || api.sessions == nil || api.projector == nil {
+		return factoryruntime.FactoryStatus{}, factoryruntime.ErrNotRunning
+	}
+	if sessionID = strings.TrimSpace(sessionID); sessionID == "" {
+		sessionID = factorysessions.DefaultSessionID
+	}
+	result, err := api.sessions.ObserveForSession(ctx, sessionID, factoryruntime.ObserveRequest{
+		Scope: factoryruntime.ObservationScopeFull,
+	})
+	if err != nil {
+		return factoryruntime.FactoryStatus{}, err
+	}
+	return api.projector.ProjectFactoryStatusFromObservation(result.Observation), nil
+}
+
 // Server is retained as a private receiver alias while the moved handler files
 // are kept mechanically identical to the established public behavior.
 type Handler = Adapter
