--- pkg/transports/mapping/composition/services.go
@@ -1,42 +1,0 @@
-package composition
-
-import (
-	factorydefinitions "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
-	factoryruntime "github.com/portpowered/infinite-you/pkg/services/factory_runtime"
-	factorysessions "github.com/portpowered/infinite-you/pkg/services/factory_sessions"
-	apisurface "github.com/portpowered/infinite-you/pkg/transports/mapping"
-	factorydefinitionmapping "github.com/portpowered/infinite-you/pkg/transports/mapping/factorydefinition"
-	factorysessionmapping "github.com/portpowered/infinite-you/pkg/transports/mapping/factorysession"
-)
-
-func NewRuntimeAPI(
-	runtime factoryruntime.Service,
-	definitions factorydefinitions.Service,
-) apisurface.RuntimeAPI {
-	return apisurface.NewRuntimeAPI(
-		runtime,
-		factorydefinitionmapping.New(definitions),
-	)
-}
-
-func NewLiveSessionAPI(
-	liveControl factorysessions.LiveControlService,
-	gateway factorysessionmapping.LiveGateway,
-) apisurface.LiveSessionAPI {
-	return factorysessionmapping.NewLiveAPI(liveControl, gateway)
-}
-
-func NewFactoryDefinitionAPI(service factorydefinitions.Service) apisurface.FactorySaveAPI {
-	definitions := factorydefinitionmapping.New(service)
-	return factorydefinitionmapping.NewAPI(definitions, definitions)
-}
-
-func NewInvocationAPI(invocations factorysessions.InvocationService) apisurface.InvocationAPI {
-	return factorysessionmapping.NewInvocationAPI(invocations)
-}
-
-func NewDurableAPI(
-	execution factorysessionmapping.DurableExecution,
-) apisurface.DurableSessionAPI {
-	return factorysessionmapping.NewDurableAPI(execution)
-}
