+++ pkg/wire/wire.go
@@ -23,11 +23,6 @@ import (
 	acp "github.com/portpowered/infinite-you/pkg/transports/acp"
 	"github.com/portpowered/infinite-you/pkg/transports/cli"
 	workersessionsrootcli "github.com/portpowered/infinite-you/pkg/transports/cli/worker_sessions"
-	httpapplication "github.com/portpowered/infinite-you/pkg/transports/http/application"
-	apisurface "github.com/portpowered/infinite-you/pkg/transports/mapping"
-	"github.com/portpowered/infinite-you/pkg/transports/mapping/composition"
-	factorydefinitionmapping "github.com/portpowered/infinite-you/pkg/transports/mapping/factorydefinition"
-	factorysessionmapping "github.com/portpowered/infinite-you/pkg/transports/mapping/factorysession"
 	mcpstdio "github.com/portpowered/infinite-you/pkg/transports/mcp/stdio"
 )
 
@@ -39,16 +34,8 @@ var platformSet = wire.NewSet(
 var apiSet = wire.NewSet(
 	providersessionshttp.NewAdapter,
 	providersessionshttp.NewHandler,
-	composition.NewHTTPBinder,
-	apisurface.NewRuntimeAPI,
-	composition.NewLiveSessionAPI,
-	factorydefinitionmapping.NewAPI,
-	factorysessionmapping.NewDurableAPI,
-	factorysessionmapping.NewLiveAPI,
-	factorysessionmapping.NewInvocationAPI,
 	mcpstdio.NewOpener,
 	provideHTTPRuntimeBinding,
-	httpapplication.NewHandler,
 )
 
 var servicesSet = wire.NewSet(
